Code generation + other stuff

+++ scanner.h
@@ -16,6 +16,7 @@
 #define __SCANNER_H__
 
 #include <stdio.h>
+#include <stdint.h>
 
 #include "error.h"
 #include "str.h"
@@ -43,7 +44,7 @@ typedef union
 {
     Keyword keyword;
     string *string;
-    int integer;
+    int64_t integer;
     double decimal;
 } Token_attribute;
